Fix Integrate RBAC buat Sidang Sempro

+++ database/migrations/2024_04_02_192928_create_users_table.php
@@ -13,27 +13,29 @@ class CreateUsersTable extends Migration
      */
     public function up()
     {
-        Schema::create('users', function (Blueprint $table) {
-            $table->id();
-            $table->uuid('public_id')->unique();
-            $table->string('name');
-            $table->string('email')->unique();
-            $table->string('nip')->nullable()->unique();
-            $table->string('nidn')->nullable()->unique();
-            $table->string('npm')->nullable()->unique();
-            $roleTable = config('permission.table_names.roles', 'roles');
-            $table->foreignId('role_id')->nullable()->constrained($roleTable)->onDelete('cascade');
-            $table->boolean('is_active')->default(false);
-            $table->string('password');
-            $table->bigInteger('phone')->nullable();
-            $table->string('location')->nullable();
-            $table->string('about_me')->nullable();
-            $table->text('image')->nullable();
-            $table->rememberToken();
-            $table->timestamps();
-            $table->timestamp('email_verified_at')->nullable();
-            $table->timestamp('deleted_at')->nullable();
-        });
+        if (!Schema::hasTable('users') && !Schema::hasTable('tb_users')) {
+            Schema::create('users', function (Blueprint $table) {
+                $table->id();
+                $table->uuid('public_id')->unique();
+                $table->string('name');
+                $table->string('email')->unique();
+                $table->string('nip')->nullable()->unique();
+                $table->string('nidn')->nullable()->unique();
+                $table->string('npm')->nullable()->unique();
+                $roleTable = config('permission.table_names.roles', 'roles');
+                $table->foreignId('role_id')->nullable()->constrained($roleTable)->onDelete('cascade');
+                $table->boolean('is_active')->default(false);
+                $table->string('password');
+                $table->bigInteger('phone')->nullable();
+                $table->string('location')->nullable();
+                $table->string('about_me')->nullable();
+                $table->text('image')->nullable();
+                $table->rememberToken();
+                $table->timestamps();
+                $table->timestamp('email_verified_at')->nullable();
+                $table->timestamp('deleted_at')->nullable();
+            });
+        }
     }
 
     /**
